Prevent discarding localization data for controls visited
This may somehow have become necessary because of the previous commit.
Without it, we delete any tu that has only the variant we are updating,
and don't re-create it. This gets called while visiting controls as a window
comes up, so all the controls we have seen are deleted from localization!