Improve LanguageChoosingDialog
Modify dialog so that all text shows on Linux and text wraps at word boundaries.
The trick is to set AutoEllipsis to true, otherwise the break will be at any
character on Mono. Also set anchors of controls so that the user can resize the
control to see all text.