Fix string extraction to work on Linux/Mono
The problem with throwing exceptions in processing Palaso.dll appears to
have gone away.  However, there is a consistent problem with looking for
and not finding resources for a large number of types, which throws
exceptions in its own right.  This doesn't seem to be critical, so a
try-catch with the specific exception allows the string extraction to work
without crashing or hanging on Linux.

These changes together fix https://jira.sil.org/browse/BL-618 "[Linux]
Missing a significant portion of entries in L10NSharp".