Fix #1 GetIsStringAvailableForLangId() doing fallback
The methods that GetIsStringAvailableForLangId() were calling had a hidden "feature", which is that if they didn't find what you asked for, they silently fell back to other languages. This of course defeats the purpose of this method. So I renamed the method that do fallback to make it obvious that they do, and had this call a lower-level method that does not.